refactor: db 관련된 로직을 OmokGameDbHelper에서 처리하도록 수정
