--- app/src/main/java/woowacourse/omok/FeedReaderDbHelper.kt
@@ -1,35 +1,0 @@
-package woowacourse.omok
-
-import android.content.Context
-import android.database.sqlite.SQLiteDatabase
-import android.database.sqlite.SQLiteOpenHelper
-import android.provider.BaseColumns
-
-class OmokGameDbHelper(context: Context) :
-    SQLiteOpenHelper(context, DATABASE_NAME, null, DATABASE_VERSION) {
-    override fun onCreate(db: SQLiteDatabase) {
-        db.execSQL(SQL_CREATE_STONE_ENTRIES)
-    }
-
-    override fun onUpgrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-        db.execSQL(SQL_DELETE_ENTRIES)
-        onCreate(db)
-    }
-
-    override fun onDowngrade(db: SQLiteDatabase, oldVersion: Int, newVersion: Int) {
-        onUpgrade(db, oldVersion, newVersion)
-    }
-
-    companion object {
-        const val DATABASE_VERSION = 1
-        const val DATABASE_NAME = "omokGame.db"
-        private const val SQL_CREATE_STONE_ENTRIES =
-            "CREATE TABLE IF NOT EXISTS ${OmokGameContract.Stone.TABLE_NAME} (" +
-                    "${BaseColumns._ID} INTEGER PRIMARY KEY," +
-                    "${OmokGameContract.Stone.COLOR} TEXT," +
-                    "${OmokGameContract.Stone.X} INTEGER," +
-                    "${OmokGameContract.Stone.Y} INTEGER)"
-        private const val SQL_DELETE_ENTRIES = "DROP TABLE IF EXISTS ${OmokGameContract.Stone.TABLE_NAME}"
-    }
-
-}
